+++ CHANGELOG.md
@@ -5,8 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased - v0.2.0
+## Unreleased - v0.3.0
+### Added
 
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## v0.2.0
 ### Added
 - `Rn2903::system_version_bytes()`
 - `Rn2903::system_factory_reset()`